requirements.txt
watchdog==0.8.3
pika==0.12.0
imutils==0.4.6
face_recognition==1.2.2
Flask==1.0.2
flask_restplus==0.12.1
pyzbar==0.1.7
Werkzeug==0.14.1
numpy==1.13.3
mysql_connector_repackaged==0.3.1
